빌드관련 오류해결
- 배포를 하기 위해 빌드를 하는 과정에서 에러가 발생
- test 과정에서 카카오 api 키를 읽지 못하는 에러였다
- build.gradle에서 test 관련 코드 추가
- 또한 쉘 스크립트에서 빌드할때 .env 파일을 우선적으로 읽게 해줬다
- 도메인 교체할 때 sql파일도 수정했어야했는데 잊고 있었다. 이 기회에 수정했다

+++ database/init/direction.sql
@@ -4,7 +4,7 @@ create table `direction` (
                              `id` bigint(20) not null auto_increment,
                              `created_date` datetime(6) default null,
                              `modified_date` datetime(6) default null,
-                             `distance` double not null,
+                             `distance` integer not null,
                              `input_address` varchar(255) collate utf8mb4_unicode_ci default null,
                              `input_latitude` double not null,
                              `input_longitude` double not null,
